Initial commit, alert renders successfully in place of app conditionally, replaced alerts with reused MCPAlerts at top level now to consolidate
